refactor methods conans->conan
